testproject/requirements.txt
Django
djangorestframework
django-environ==0.4.5
django-cors-headers==3.7.0
djangorestframework-simplejwt>=4.8.0
pyotp>=2.6.0
twilio>=7.0.0
yubico-client>=1.13.0
smsapi-client>=2.4.5
pyjwt<=2.0.1

drf-yasg==1.20.0

pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.1.0
attrs==20.3.0
coverage==5.5

isort
flake8
